+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added support for querying inverse relationships (#2904).
 * Moved inverse relationships out of beta stage.
+* Added `Realm.getDefaultConfiguration()` (#4725).
 
 ### Bug Fixes